Update `Display for SyntaxShape` to `float`
Separate commit as this may affect tooling if we have a location where
`SyntaxShape::to_string` or similar is read.

+++ crates/nu-protocol/src/syntax_shape.rs
@@ -189,7 +189,7 @@ impl Display for SyntaxShape {
             SyntaxShape::Number => write!(f, "number"),
             SyntaxShape::Range => write!(f, "range"),
             SyntaxShape::Int => write!(f, "int"),
-            SyntaxShape::Float => write!(f, "decimal"),
+            SyntaxShape::Float => write!(f, "float"),
             SyntaxShape::Filepath => write!(f, "path"),
             SyntaxShape::Directory => write!(f, "directory"),
             SyntaxShape::GlobPattern => write!(f, "glob"),
